# docs/openapi/admin-v1.yaml
openapi: 3.1.0
info:
  title: Hearts by duanap Admin API
  version: 1.0.0
  description: 同源后台管理接口。除登录外均使用管理会话 Cookie；写操作同时要求 CSRF 和幂等键。
servers:
  - url: /api/admin/v1
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getAdminSession
      responses:
        '200':
          description: 当前管理员会话。
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthEnvelope' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: adminLogout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: hearts_admin_sid
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...
  schemas:
    BaseEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string }
...
    AuthEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [admin, csrfToken, expiresAt]
              properties:
                admin: { $ref: '#/components/schemas/Admin' }
                csrfToken: { type: string }
                expiresAt: { type: string, format: date-time }
    Admin:
      type: object
      required: [adminId, username, role, active]
      properties:
        adminId: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [super_admin, operator, viewer] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId: { type: string }
  responses:
    Success:
      description: 请求成功。
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BaseEnvelope' }
    Error:
      description: 标准错误响应。
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
security:
  - adminSession: []